+++ R/ImportData.R
@@ -3,54 +3,62 @@
 #' @description Imports collected data from file into a dataframe of class \code{datasource} and specified 
 #' \code{socialmedia} type that is usable by \code{\link{Create}} functions. 
 #' 
-#' @param file Character string. Collected data file path.
-#' @param type Character string. Type of file or file format of file to import \code{csv} or \code{rds}. Default is 
-#' \code{csv}.
+#' @param path Character string. Collected data file path.
 #' @param socialmedia Character string. Social media type of collected data \code{twitter}, \code{youtube} or 
 #' \code{reddit}.
+#' @param type Character string. Type of file or file format of file to import \code{csv} or \code{rds}. Default is 
+#' \code{NULL} to use extension.
 #' 
 #' @return A dataframe with datasource class attributes.
 #' 
 #' @aliases ImportData
 #' @name vosonSML::ImportData
 #' @export
-ImportData <- function(file, type = "csv", socialmedia) {
+ImportData <- function(path, socialmedia, type = NULL) {
   
-  if (missing(file)) {
-    stop("Please provide file path of data to import.", .call = FALSE) 
+  # expected import types
+  supported_types <- c("csv", "rds")
+  
+  if (missing(path)) {
+    stop("Please provide file path of data to import.", call. = FALSE) 
   }
   
   if (missing(socialmedia)) {
-    stop("Please provide the social media type of data to import.", .call = FALSE) 
+    stop("Please provide the social media type of data to import.", call. = FALSE) 
   }
   
-  supported_types <- c("csv", "rds")
+  # if type is null check file extension
+  if (is.null(type)) {
+    type <- gsub(".*\\.([A-Za-z]{3})$", "\\1", path, ignore.case = TRUE, perl = TRUE)
+  }
   
+  type <- tolower(trimws(type))
   if (!type %in% supported_types) {
-    err <- paste0("File format not supported. please choose from: ", paste0(supported_types, collapse = ", "), ".")
-    stop(err, .call = FALSE) 
+    stop(paste0("File format not supported. please choose from: ", paste0(supported_types, collapse = ", "), "."), 
+         call. = FALSE) 
   }
   
   df <- tryCatch({
     switch(type,
-      "csv" = read.csv(file),
-      "rds" = readRDS(file = file))
+      "csv" = read.csv(path),
+      "rds" = readRDS(file = path))
     
   }, error = function(e) {
-    err <- paste0("Could not read: ", file, ".\n", gsub("^Error:\\s", "", paste0(e)))
-    stop(err, .call = FALSE)
+    stop(paste0("Could not read: ", path, ".\n", gsub("^Error:\\s", "", paste0(e))), call. = FALSE)
   })
   
-  cat(paste0(toupper(type), " file read: ", file))
+  cat(paste0(toupper(type), " file read: ", path), "\n")
   
-  class(df) <- c("data.table", "data.frame", "datasource")
+  if (!is.data.frame(df)) {
+    stop("Read data is not a dataframe.", call. = FALSE)
+  }
   
-  socialmedia <- tolower(socialmedia)
-  switch(socialmedia,
-    twitter = { class(df) <- append(class(df), c("twitter")) },
-    youtube = { class(df) <- append(class(df), c("youtube")) },
-    reddit = { class(df) <- append(class(df), c("reddit")) },
-    stop("Unknown social media type provided as datasource.", .call = FALSE))
+  class(df) <- union(class(df), c("datasource"))
+  class(df) <-switch(tolower(trimws(socialmedia)),
+    twitter = { union(class(df), c("twitter")) },
+    youtube = { union(class(df), c("youtube")) },
+    reddit = { union(class(df), c("reddit")) },
+    stop("Unknown social media type provided as datasource.", call. = FALSE))
   
-  return(df)
+  df
 }