+++ NEWS.md
@@ -1,9 +1,8 @@
 # vosonSML 0.29.5
 
 ## Minor Changes
-- Removed the `rtweet` and `tictoc` packages from dependency imports to suggested packages. Using
-  the `Authenticate` and `Collect` functions for `twitter` without the `rtweet` package installed
-  will stop and notify the user that it must be installed first.
+- Removed the `tictoc` package from dependency imports to suggested packages. Added some checks
+  for whether the `rtweet` package is installed.
 
 # vosonSML 0.29.4